requirements.txt
fastapi
pydantic
uvicorn
python-kubernetes
kfp
